requirements.txt
aio-pika
aiormq
annotated-types
anyio
certifi
charset-normalizer
click
colorlog
dnspython
email_validator
exceptiongroup
fastapi
fastapi-cli
h11
httpcore
httptools
httpx
idna
Jinja2
markdown-it-py
MarkupSafe
mdurl
mistune
multidict
pamqp
propcache
pydantic
pydantic_core
Pygments
python-dotenv
python-multipart
PyYAML
requests
rich
rich-toolkit
shellingham
sniffio
starlette
typer
typing_extensions
urllib3
uvicorn
uvloop
watchfiles
websockets
yarl
